part4_21

+++ part4/bloglist/controllers/blogs.js
@@ -56,8 +56,22 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user.toString() === user._id.toString()) {    
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else{
+    return response.status(401).json({ error: 'only the user who created the blog can delete it' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
@@ -67,11 +81,14 @@ blogsRouter.put('/:id', async (request, response, next) => {
     body.likes = 0
   }
 
+  const blogToUpdate = await Blog.findById(request.params.id)
+
   const blog = {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user : blogToUpdate.user
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
